Improve CNN model
The architecture of the CNN should now completely match the one as
described in the original publication. In addition, it now also has
the option to apply batch normalization in every convolutional block.
Not sure whether this will improve the results, but it might be worth
it to give it a try.